pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "backup-tool"
version = "0.1.0"
description = "Backup directory snapshots to Google Drive"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv",
    "google-auth",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "google-api-python-client",
]

[project.optional-dependencies]
test = ["pytest>=7"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["backup_tool*"]

[tool.pytest.ini_options]
testpaths = ["src/backup_tool/tests"]
